Add Warp Plus Key generation application
A Warp Plus Key Generator application is added in this commit. It also includes a README providing an explanation of the application functionality, usage guide and necessary prerequisites. Further it involves key generation script and associated requirements.txt file specifying dependencies. The creation of new .gitignore file specifies which files Git should ignore, and the addition of MIT License is also done for the project.